requirements.txt
pytest
Appium-Python-Client
requests
